refactor Tour and Tooltip Item
